Extract hook install helpers
